pyproject.toml
[project]
name = "f1-broadcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "line-bot-sdk>=3.19.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "schedule>=1.2.2",
    "uvicorn>=0.35.0",
]
